Refactor component testing boilerplate out into its own file
